+++ lib/td_dd_web/controllers/fallback_controller.ex
@@ -31,6 +31,10 @@ defmodule TdDdWeb.FallbackController do
     |> send_resp(:not_found, Jason.encode!(%{message: message}))
   end
 
+  def call(conn, nil) do
+    render_error(conn, :not_found)
+  end
+
   def call(conn, {:error, :not_found}) do
     render_error(conn, :not_found)
   end